Merged

+++ iOS/G3MiOSSDK/Commons/Mesh/CompositeMesh.cpp
@@ -109,9 +109,9 @@ void CompositeMesh::zRawRender(const G3MRenderContext* rc, const GLState* parent
   }
 }
 
-void CompositeMesh::showNormals(bool v) const{
-  const int childrenCount = _children.size();
-  for (int i = 0; i < childrenCount; i++) {
+void CompositeMesh::showNormals(bool v) const {
+  const size_t childrenCount = _children.size();
+  for (size_t i = 0; i < childrenCount; i++) {
     Mesh* child = _children[i];
     child->showNormals(v);
   }
