+++ iOS/G3MiOSSDK/Commons/Rendererers/PlanetRenderer.cpp
@@ -223,9 +223,9 @@ PlanetRenderer::~PlanetRenderer() {
   delete _lastSplitTimer;
   
   delete _lastVisibleSector;
-  
-  const int visibleSectorListenersCount = _visibleSectorListeners.size();
-  for (int i = 0; i < visibleSectorListenersCount; i++) {
+
+  const size_t visibleSectorListenersCount = _visibleSectorListeners.size();
+  for (size_t i = 0; i < visibleSectorListenersCount; i++) {
     VisibleSectorListenerEntry* entry = _visibleSectorListeners[i];
     delete entry;
   }
@@ -342,9 +342,9 @@ void PlanetRenderer::createFirstLevelTiles(std::vector<Tile*>& firstLevelTiles,
     std::vector<Tile*>* children = tile->createSubTiles(splitLatitude,
                                                         splitLongitude,
                                                         false);
-    
-    const int childrenSize = children->size();
-    for (int i = 0; i < childrenSize; i++) {
+
+    const size_t childrenSize = children->size();
+    for (size_t i = 0; i < childrenSize; i++) {
       Tile* child = children->at(i);
       createFirstLevelTiles(firstLevelTiles, child, firstLevel);
     }
@@ -488,12 +488,12 @@ RenderState PlanetRenderer::getRenderState(const G3MRenderContext* rc) {
   //CALLING THIS UNTIL TEXTURE AND ELEV. ARE SOLVED AS TEXTURE IS REQUESTED AFTER ED IS RESOLVED
   if (!_allFirstLevelTilesAreTextureSolved) {
     
-    const int firstLevelTilesCount = _firstLevelTiles.size();
+    const size_t firstLevelTilesCount = _firstLevelTiles.size();
     
     if (_tilesRenderParameters->_forceFirstLevelTilesRenderOnStart) {
       _statistics.clear();
-      
-      for (int i = 0; i < firstLevelTilesCount; i++) {
+
+      for (size_t i = 0; i < firstLevelTilesCount; i++) {
         Tile* tile = _firstLevelTiles[i];
         tile->prepareForFullRendering(rc,
                                       _texturizer,
@@ -566,9 +566,9 @@ void PlanetRenderer::visitTilesTouchesWith(const Sector& sector,
         layers.push_back(layer);
       }
     }
-    
-    const int firstLevelTilesCount = _firstLevelTiles.size();
-    for (int i = 0; i < firstLevelTilesCount; i++) {
+
+    const size_t firstLevelTilesCount = _firstLevelTiles.size();
+    for (size_t i = 0; i < firstLevelTilesCount; i++) {
       Tile* tile = _firstLevelTiles[i];
       if (tile->_sector.touchesWith(sector)) {
         _tileVisitor->visitTile(layers, tile);
@@ -592,9 +592,9 @@ void PlanetRenderer::visitSubTilesTouchesWith(std::vector<Layer*> layers,
                                               const int maxLevel) {
   if (tile->_level < maxLevel) {
     std::vector<Tile*>* subTiles = tile->getSubTiles();
-    
-    const int subTilesCount = subTiles->size();
-    for (int i = 0; i < subTilesCount; i++) {
+
+    const size_t subTilesCount = subTiles->size();
+    for (size_t i = 0; i < subTilesCount; i++) {
       Tile* tl = subTiles->at(i);
       if (tl->_sector.touchesWith(sectorToVisit)) {
         if ((tile->_level >= topLevel)) {
@@ -689,8 +689,8 @@ void PlanetRenderer::render(const G3MRenderContext* rc,
   _lastVisibleSector = _statistics.updateVisibleSector(_lastVisibleSector);
   // ILogger::instance()->logInfo("=> visibleSector: %s", _lastVisibleSector->description().c_str());
   if (_lastVisibleSector != NULL) {
-    const int visibleSectorListenersCount = _visibleSectorListeners.size();
-    for (int i = 0; i < visibleSectorListenersCount; i++) {
+    const size_t visibleSectorListenersCount = _visibleSectorListeners.size();
+    for (size_t i = 0; i < visibleSectorListenersCount; i++) {
       VisibleSectorListenerEntry* entry = _visibleSectorListeners[i];
       entry->tryToNotifyListener(_lastVisibleSector, rc);
     }
@@ -726,7 +726,7 @@ bool PlanetRenderer::onTouchEvent(const G3MEventContext* ec,
     
     const Geodetic3D position = planet->toGeodetic3D(positionCartesian);
     
-    const int firstLevelTilesCount = _firstLevelTiles.size();
+    const size_t firstLevelTilesCount = _firstLevelTiles.size();
     for (int i = 0; i < firstLevelTilesCount; i++) {
       const Tile* tile = _firstLevelTiles[i]->getDeepestTileContaining(position);
       if (tile != NULL) {
